% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{clean_merges}
\alias{clean_merges}
\title{Cleans the merges dataframe to be compatible with cutree function}
\usage{
clean_merges(merges, n)
}
\arguments{
\item{merges}{Dataframe of cluster merges by step}

\item{n}{Number of rows in original dataframe}
}
\value{
A dataframe of cleaned merges by step
}
\description{
Cleans the merges dataframe to be compatible with cutree function
}
