% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridgereg_coef.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Coefficient}
\usage{
\method{coef}{ridgereg}(x, ...)
}
\arguments{
\item{x}{an object of class.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
coefficients extracted from the model object.
}
\description{
\code{coef} extract model coefficients from objects of class \code{"ridgereg"}.
}

