% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createvector.R
\name{createvector}
\alias{createvector}
\title{Creates Vector from 3 Objects}
\usage{
createvector(a, b, c)
}
\arguments{
\item{a}{object A}

\item{b}{object B}

\item{c}{object C}
}
\value{
a new numeric vectore binding A, B, and C
}
\description{
This function takes 3 objects of data type character, numeric, integer or factor
and concatenates them into a numeric vector.
}
\examples{

createvector(factor(c(1,2)),c(3,4),c("5","6"))

}
