% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_ridgereg.R
\docType{data}
\name{cridgereg}
\alias{cridgereg}
\title{Ridge Regression caret model}
\format{An object of class \code{list} of length 10.}
\usage{
cridgereg
}
\description{
A caret model of the ridge regression
}
\keyword{datasets}

