% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{Clusters observations into groups using kmeans clustering}
\usage{
k_means(dat, k)
}
\arguments{
\item{dat}{A Data Frame}
}
\value{
A list of cluster assignments and the total sum of squares
}
\description{
Clusters observations into groups using kmeans clustering
}
