\name{lab7-package}
\alias{lab7-package}
\alias{lab7}
\docType{package}
\title{
\packageTitle{lab7}
}
\description{
\packageDescription{lab7}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{lab7}
\packageIndices{lab7}
}
\author{
\packageAuthor{lab7}

Maintainer: \packageMaintainer{lab7}
}
\keyword{ package }

