% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lab7.R
\docType{package}
\name{Lab7}
\alias{Lab7}
\alias{Lab7-package}
\title{Lab7 package contains a RC generator for ridge regression analysis and some
data munging}
\description{
The Lab7 package consists of a single RC generator and some data munging.
Package imports packages \code{ggplot2}, \code{nycflights13}, \code{mlbench},
\code{dplyr}, \code{caret}, \code{grid}, \code{gridExtra}.
}
\section{Lab7 functions}{
 The Lab7 function \code{ridgereg} creates an
 \code{ridgereg} class object (actually an environment in R) which does ridge
 regression analysis upon input formula and data.frame and parameter lambda.
 All the calculations are done and stored in the fields of the object upon
 initialization. The methods just returns the values.
}

