% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{LinearRegression}
\description{
Returns the result of the Linear Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}
}}


