% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear regression class}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{Formula of the regression}

\item{data}{Data set used}
}
\description{
Calculate linear regression from a dataset
}
