% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{predict ridgereg}
\usage{
\method{predict}{ridgereg}(x)
}
\arguments{
\item{x}{linreg class object}
}
\description{
Return the fitted values of the ridgeear regression
}
