% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{print ridgereg}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{linreg class object}

\item{...}{other parameters}
}
\description{
Print the ridgereg class
}
