% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{print.ridgereg1}
\alias{print.ridgereg1}
\title{Print a ridgereg object}
\usage{
\method{print}{ridgereg1}(x)
}
\arguments{
\item{x}{A ridgereg object.}
}
\value{
coefficients.
}
\description{
Print a ridgereg object
}
\examples{
a <- ridgereg1(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda=0)
print(a)
}

