% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Title Ridge Regression}
\value{
returns regression coeffiecient beta hat and fitted values.
}
\description{
Title Ridge Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{parsedata}}{character.}

\item{\code{identity_matrix}}{matrix.}

\item{\code{beta_ridge_QR}}{matrix.}

\item{\code{ybar_ridge_QR}}{matrix.}
}}


\examples{
 data(iris)
}
