% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Perform ridge regression using QR decomposition}
\usage{
ridgereg(formula, data, subset, na.action, lambda = 0, model = FALSE,
  x = FALSE, y = FALSE, contrasts = NULL, ...)
}
\arguments{
\item{formula}{an object of \code{\link{class}} \code{\link{formula}} (or one
that can be coerced to that class): a symbolic description of the model to
be fitted.}

\item{data}{a data frame or (or object coercible by
\code{\link{as.data.frame}}).}

\item{lambda}{the lambda of the data set. Default is 0.}
}
\value{
\code{ridgereg} returns an object of class "\code{Ridgereg}".
}
\description{
Use \code{ridgereg} to perform ridge regression.
}
