% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{ridgereg1}
\alias{ridgereg1}
\title{A function for ridge regression.}
\usage{
ridgereg1(formula, data, lambda = 0)
}
\arguments{
\item{formula}{A formula (e.g. y ~ x)}

\item{data}{data frame}

\item{lambda}{lambda}
}
\value{
ridgereg object.
}
\description{
A function for ridge regression.
}
\examples{
a <- ridgereg1(Sepal.Length ~ Sepal.Width + Petal.Length, iris, 0)
}

