% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_QR.R
\docType{class}
\name{ridgereg_QR-class}
\alias{ridgereg_QR-class}
\alias{ridgereg_QR}
\title{Ridge Regression - QR decomposition method}
\value{
returns regression coeffiecient beta hat and fitted values.
}
\description{
Ridge Regression - QR decomposition method
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{lambda}}{numeric.}

\item{\code{data_name}}{character.}

\item{\code{identity_matrix}}{matrix.}

\item{\code{beta_hat_ridge_QR}}{matrix.}

\item{\code{y_hat_ridge_QR}}{matrix.}
}}


