% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression_model.R
\docType{class}
\name{RidgeRegressionModel-class}
\alias{RidgeRegressionModel}
\alias{RidgeRegressionModel-class}
\alias{ridgeregmod}
\title{A Reference Class to represent Ridge regression models.}
\description{
Stores information about a Ridge regression model generated by \code{ridgereg}.
}
\section{Fields}{

\describe{
\item{\code{call}}{Object of class \code{"character"}. A string representation of the calling code.}

\item{\code{coefficients}}{Object of class \code{"list"}. A list containing information about the estimated coefficients.
\describe{
\item{\code{val}:}{The estimated coefficients.}
\item{\code{se}:}{The standard errors of the coefficients.}
\item{\code{var}:}{The variance of the coefficients.}
\item{\code{tval}:}{The t values of the the coefficients.}
\item{\code{pval}:}{The p values of the coefficients.}
}}

\item{\code{fitted.values}}{The fitted values of y using the model's coefficients.}

\item{\code{df}}{The number of degrees of freedom in the model.}
}}
\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the estimated coefficients.}

\item{\code{pred()}}{Returns the fitted values.}

\item{\code{print()}}{Prints the model.}

\item{\code{summary()}}{Shows a summary of the model.}
}}

