% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{theme_liu}
\alias{theme_liu}
\title{theme function}
\usage{
theme_liu(base_size = 11, base_family = "sans")
}
\arguments{
\item{base_size}{size of base}

\item{base_family}{font}
}
\description{
Theme for LIU plots
}
