% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_ridge.R
\name{try_ridge}
\alias{try_ridge}
\title{using Caret Ridge BostonHousing}
\usage{
try_ridge(X)
}
\arguments{
\item{BostonHousing}{only works with this data set at the moment.}
}
\value{
\code{try_ridge} returns an object of class "\code{ridgereg}".
}
\description{
Use \code{try_ridge} to perform ridge regression.
}
\seealso{
\code{\link{lm}}, \code{\link{class}}, \code{\link{formula}}
}
