% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{clean_merges}
\alias{clean_merges}
\title{Edits the Merges Dataframe to be Compatible with cutree()}
\usage{
clean_merges(merges, n)
}
\arguments{
\item{merges}{A dataframe of the merges at each step}
}
\value{
A dataframe of merges at each step with positive
numbers representing a merge with a cluster formed at that row
}
\description{
Edits the Merges Dataframe to be Compatible with cutree()
}
