% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{hier_clust}
\alias{hier_clust}
\title{Clusters observations into groups using agglomerative hierarchical clustering}
\usage{
hier_clust(dat, k)
}
\arguments{
\item{dat}{A Data Frame}
}
\value{
A Vector of Cluster Assignments
}
\description{
Clusters observations into groups using agglomerative hierarchical clustering
}
