\name{lab7-package}
\alias{lab7-package}
\alias{lab7}
\docType{package}
\title{
\packageTitle{lab7}
}
\description{
\packageDescription{lab7}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab7}
\packageIndices{lab7}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lab7}

Maintainer: \packageMaintainer{lab7}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
