% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridgereg_predict.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict method for Ridge regression model}
\usage{
\method{predict}{ridgereg}(x, newdata = NULL)
}
\arguments{
\item{x}{an object of class.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
produce a vector of predictions.
}
\description{
\code{predict} give predict values based on class \code{"ridgereg"}.
}

