% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridgereg_print.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print values}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
print out the coefficients and coefficient names.
}
\description{
\code{print} prints its argument and returns it out for class \code{"ridgereg"}.
}

