% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{A RC class to use ridge regression}
\description{
A RC class to use ridge regression
}
\details{
Please look at the vignette.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula}

\item{\code{data}}{data frame}

\item{\code{lambda}}{lambda}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns ridge regression regression coefficients}

\item{\code{predict(datanew = NULL)}}{Returns predicted values}

\item{\code{print()}}{Prints the input and the coefficients in a user-friedly way}
}}

\examples{
r <- ridgereg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris,lambda=2)
}
