% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{A RC class to use ridge regression}
\description{
A RC class to use ridge regression
}
\details{
Please look at the vignette.
}
\section{Fields}{

\describe{
\item{\code{}}{}
}}

\section{Methods}{

\describe{
\item{\code{print()}}{Prints the input and the coefficients in a user-friedly way}
}}

\examples{
my_address<-address$new("Linkoping university")
}
