% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{A RC class for ridge regression models}
\description{
A RC class for ridge regression models
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{data}}{A dataset which we apply the formula (data.frame)}

\item{\code{lambda}}{A numeric parameter}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the regression coefficients}

\item{\code{pred()}}{Returns fitted values}

\item{\code{print()}}{Prints the call of the class and ridge regression coefficients in named vector}

\item{\code{resid()}}{Returns residual values}
}}

