% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression}
\description{
Returns the result of the ridge regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula}

\item{\code{data}}{data.frame}
}}


