% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression}
\value{
Regression coefficent beta hat and fitted values.
}
\description{
Ridge Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{lambda}}{numeric.}

\item{\code{beta_hat_ridge}}{matrix.}

\item{\code{y_hat}}{numeric.}

\item{\code{coef_ridge}}{numeric.}

\item{\code{data_name}}{character.}
}}


