% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridgereg.r
\docType{class}
\name{ridgereg-class}
\alias{ridgereg}
\alias{ridgereg-class}
\title{Function for ridge regression}
\arguments{
\item{formula}{formula.}

\item{data}{data.frame.}

\item{lambda}{numeric.}
}
\value{
A linreg object.
}
\description{
Function for ridge regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula.}

\item{\code{data}}{A data.frame.}

\item{\code{bhat}}{A matrix.}

\item{\code{yhat}}{A matrix.}
}}
\examples{
ridgereg1 <- ridgereg(formula = Petal.Length ~ Species, data = iris, lambda = 0)
ridgereg1$print()
ridgereg1$coef()
ridgereg1$pred()
}

