% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge regression}
\usage{
ridgereg(formula, data, lambda = 0.1)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{lambda}{an optional parameter which is a scalar of ridge constant. The default value is set to 0.1.}

\item{waiting}{an optional data frame or list containing the variables in the model.}
}
\value{
Returns an object of class \code{"ridgereg"}

Followings are object of class \code{"ridgereg"}:
\itemize{
 \item \code{coefficients}: a named vector of ridge regression coefficients
 \item \code{predict}: a predicted values
}
}
\description{
It is one type of regression using to reduce multicollinearity amongst regression predictor variables in a model.
}
\examples{
ridereg(eruptions~waiting, data=faithful)
}

