% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R, R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge Regression}
\source{
\url{https://en.wikipedia.org/wiki/Linear_regression}

\url{https://en.wikipedia.org/wiki/Linear_regression}
}
\usage{
ridgereg(formula, data, lambda=0)

ridgereg(formula, data, lambda=0)
}
\arguments{
\item{formula}{an object of class \code{formula}. Describes the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{lambda}{A constant >= 0}

\item{formula}{an object of class \code{formula}. Describes the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{lambda}{A constant >= 0}
}
\value{
The fitted model of class \code{"LinearRegressionModel"}.

The fitted model of class \code{"LinearRegressionModel"}.
}
\description{
ridgereg is used to fit ridge regression models.

ridgereg is used to fit ridge regression models.
}
\examples{
ridgereg(formula = Petal.Width ~ Petal.Length, data=iris)
ridgereg(formula = Petal.Width ~ Petal.Length, data=iris, lambda=2)

ridgereg(formula = Petal.Width ~ Petal.Length, data=iris)
ridgereg(formula = Petal.Width ~ Petal.Length, data=iris, lambda=2)

}

