% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg1}
\alias{ridgereg1}
\title{Ridge regression}
\usage{
ridgereg1(y, x, lambda = 0.1)
}
\arguments{
\item{lambda}{an optional parameter which is a scalar of ridge constant. The default value is set to 0.1.}

\item{Y}{is the matrix of dependent variable}

\item{X}{independent variable}
}
\value{
Returns an object of class \code{"ridgereg"}

Followings are object of class \code{"ridgereg"}:
\itemize{
 \item \code{coefficients}: a named vector of ridge regression coefficients
 \item \code{predict}: a predicted values
}
}
\description{
It is one type of regression using to reduce multicollinearity amongst regression predictor variables in a model.
Rigreg1 does not take a formula but two matrizes.
}
\examples{
ridereg(eruptions~waiting, data=faithful)
}

