% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression_model.R
\docType{class}
\name{RidgeRegressionModel-class}
\alias{RidgeRegressionModel}
\alias{RidgeRegressionModel-class}
\alias{ridgeregmod}
\title{A Reference Class to represent Ridge regression models.}
\description{
Stores information about a Ridge regression model generated by \code{ridgereg}.
}
\section{Fields}{

\describe{
\item{\code{call}}{Object of class \code{"character"}. A string representation of the calling code.}

\item{\code{coefficients}}{Object of class \code{"numeric"}. Contains the coefficients.}

\item{\code{fitted.values}}{The fitted values of y using the model's coefficients.}

\item{\code{center}}{The values used to center the data.}

\item{\code{scale}}{The values used to scale the data.}

\item{\code{variables}}{The variables used to create the model.}

\item{\code{lambda}}{The lambda parameter, >= 0.}
}}
\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the estimated coefficients.}

\item{\code{pred(X = NULL)}}{Returns the lastest fitted values or
if X is specified it will return new predicted values.
X do not need to be scaled, it is performed inside this method.}

\item{\code{print()}}{Prints the model.}
}}

