% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize airport delays}
\usage{
visualize_airport_delays()
}
\description{
Visualize airport delays
}
\details{
Plot of geographical position of airports of the "flights" dataset with their mean delay.
}
