% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{A function which data munges nycflights13 and plots the mean arrival delays
of airports.}
\usage{
visualize_airport_delays()
}
\value{
p A ggplot object of mean arrival delays of flights from New York to
  cities outside NYC.
}
\description{
This function data munges nycflights13 and generates a plot with mean arrival
delays at different airports.
}

