% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{This function creates a plot that visualizes the mean delay of fights 
for different airports by longitude and latitude}
\usage{
visualize_airport_delays()
}
\value{
A ggplot2
}
\description{
This function creates a plot that visualizes the mean delay of fights 
for different airports by longitude and latitude
}
