% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Handling large datasets with dplyr}
\usage{
visualize_airport_delays()
}
\value{
returns plot.
}
\description{
This function creates a plot which visualizes a large dataset.
}
\examples{
visualize_airport_delays()

}
\references{
\href{https://www.rstudio.com/wp-content/uploads/2015/02/data-wrangling-cheatsheet.pdf}{DataWrangling}
}
